moved all components out of NavbaWeb and into a parent components folder.
