+++ src/containers/navbar/components/NavbarUpper/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import NavbarLanguage from './navbar_web_language/';
+import NavbarLanguage from '../navbar_web_language/';
 import NavbarUserActions from './navbar_web_userActions/';
 import NavbarCart from './navbar_web_cart/container/';
