+++ src/containers/navbar/components/NavbarWebNavs/index.js
@@ -1,0 +1,3 @@
+export NavbarNavInfo from './navbarNavs_info';
+export NavbarNavJuices from './navbarNavs_juices';
+export NavbarNavMedia from './navbarNavs_media';
