+++ src/containers/navbar/components/index.js
@@ -1,3 +1,5 @@
-export NavbarLogo from './NavbarLogo';
+export NavbarWebLogo from './NavbarWebLogo';
 export NavbarMobile from './NavbarMobile';
 export NavbarWeb from './NavbarWeb';
+export NavbarLower from './NavbarLower';
+export NavbarUpper from './NavbarUpper';
